Refactor BaseNDManager
